@@ -5,19 +5,18 @@
                     <div class="col-lg-12">
                         <div class="card card-outline-primary">
                             <div class="card-header">
-                                <h4 class="m-b-0 text-white">USA SMS Sender</h4>
+                                <h4 class="m-b-0 text-white">SMS Sender</h4>
                             </div>
                             <div class="card-body">
-                                <form action="{{url('sendd')}}" method="post" enctype="multipart/form-data">
-								   <input type="hidden" id="uuu" name="_token" value="{{csrf_token()}}">
-                                    <div class="form-body">
+                                <form action="#">
+								    <div class="form-body">
                                         <h3 class="card-title m-t-15">Content</h3>
                                         <hr>
 										<div class="row p-t-20">
 										<div class="col-md-12">
                                                 <div class="form-group">
                                                     <label class="control-label">From</label>
-                                                    <input type="text" id="from" name="from" class="form-control" placeholder="From" value="+14243061051" readonly>
+                                                    <input type="text" id="from" name="from" class="form-control" placeholder="From" value="My Name" readonly>
                                                     <small id="error-from" class="form-control-feedback text-danger"> This field is required. </small>  
 											    </div><br>		
                                             </div>
@@ -41,7 +40,7 @@
                                                     <h4 class="card-title">Recipients</h4>
                                                     <h6 class="card-subtitle">Paste your recipient phone numbers here</h6>
                                                         <div class="form-group">
-                                                            <textarea  id="to" name="to" class="form-control" rows="15" placeholder="Enter numbers here (one per line, each number must start with +1)" style="height:450px"></textarea>
+                                                            <textarea  id="to" name="to" class="form-control" rows="15" placeholder="Enter numbers here (one per line)" style="height:450px"></textarea>
                                                            <small id="error-to" class="form-control-feedback text-danger"> This field is required. </small>
                                                         </div>
                                                 </div>
